moving task related types to task.type.ts
