+++ src/shared/types/api.type.ts
@@ -1,5 +1,3 @@
-import { TaskProps } from './domains/task.type';
-
 export type CrudOperation = 'create' | 'read' | 'update' | 'delete' | 'toggle';
 export type EntityType = 'task' | 'user' | 'setting';
 
@@ -34,50 +32,6 @@ export type PaginationParams = {
   page: number;
   limit: number;
   _t?: number; // Optional delay for testing purposes
-};
-
-export type TaskStatsProps = {
-  total: number;
-  completed: number;
-  lastUpdated?: string;
-};
-
-export type TaskBoardState = {
-  tasks: TaskProps[];
-  isLoading: boolean;
-  error: Error | null;
-  pagination: PaginationMetadata;
-  taskStats: TaskStatsProps;
-  refetch?: () => void;
-  setCurrentPage: (page: number) => void;
-  ref?: React.RefObject<HTMLElement>;
-};
-
-export type TaskBoardPresenterProps = {
-  // Data
-  tasks: TaskProps[];
-  pagination: {
-    currentPage: number;
-    totalPages: number;
-    totalCount: number;
-    completedCount?: number;
-    isFirstPage: boolean;
-    isLastPage: boolean;
-    hasMore: boolean;
-    hasPrev: boolean;
-  };
-
-  // Loading states
-  isLoading: boolean;
-  isAdding?: boolean;
-  isUpdating?: boolean;
-
-  // Error handling
-  error?: Error;
-
-  // Callbacks
-  onPageChange: (page: number) => void;
-  onClose: () => void;
 };
 
 export type OffsetPaginationProps = {
