+++ src/shared/types/domains/task.type.ts
@@ -45,3 +45,47 @@ export type UseUpdateTaskProps = Pick<
 > & {
   onSuccess?: () => void;
 };
+
+export type TaskStatsProps = {
+  total: number;
+  completed: number;
+  lastUpdated?: string;
+};
+
+export type TaskBoardState = {
+  tasks: TaskProps[];
+  isLoading: boolean;
+  error: Error | null;
+  pagination: PaginationMetadata;
+  taskStats: TaskStatsProps;
+  refetch?: () => void;
+  setCurrentPage: (page: number) => void;
+  ref?: React.RefObject<HTMLElement>;
+};
+
+export type TaskBoardPresenterProps = {
+  // Data
+  tasks: TaskProps[];
+  pagination: {
+    currentPage: number;
+    totalPages: number;
+    totalCount: number;
+    completedCount?: number;
+    isFirstPage: boolean;
+    isLastPage: boolean;
+    hasMore: boolean;
+    hasPrev: boolean;
+  };
+
+  // Loading states
+  isLoading: boolean;
+  isAdding?: boolean;
+  isUpdating?: boolean;
+
+  // Error handling
+  error?: Error;
+
+  // Callbacks
+  onPageChange: (page: number) => void;
+  onClose: () => void;
+};
